requirements.txt
torch==2.0
tensorboard==2.3
numpy
scipy
open3d>=0.8
Pillow
tqdm
